requirements.txt
flask
torch
torchvision
transformers
pillow
requests
scikit-learn